Set all module types to be stored in a single list
All modules are now stored in a single list as the communications was
easier to do in a registry/message style.  This allows for easy
implementation of different module types.  Just make a module and
register it so it can listen to data and post data.  However, there
should not be two modules posting to the same output tag.  If a new one
is going to be registered the old one needs to be unbound first.