Added ConsoleLogger and implemented module priority.
The module priority will allow for the modules to be sorted when the
registry loop is started so that modules will operate in a definitive
order.  This is helpful so that some preprocessing modules can be
selected that can condition some data and rebroadcast it under the same
Inter-Module Variable tag.

priority 0-25 should be collecting data from sensors and user inputs

priority 25-50 should be used for data preprocessing

priority 50-75 should be used for primary processing and decision making

priority 75-100 should be used for outputs

Multiple modules can share the same priority but care should be taken
that this is not done with modules that are dependent on each other as
the order of modules with the same priority is first found first served.

ConsoleLogger represents a module that is agnostic to priority as its
purpose is not to "do" anything but take status and log messages and
display them to the Console as it receives them.