Modules now have a startup and shutdown feature
The startup and shutdown functions are called before the registry begins
to process modules and after it has finished its last cycle of
processing modules, respectively.  This is to be used for opening
communications with external components or other features that otherwise
may end up running when the system is idle.