Removed the close/minimize/maximize bar
If the closes the form closes abnormally problems can arise.  It is
safer to set the WindowStyle->ControlBox property to false and disable
the close button.  You can re-enable it if you make sure that it behaves
as if the Ok button was clicked (ie it calls the setDataToModule
function.  Though going through the Ok button (and possible a Cancel
button if the user made a mistake) prevents the user from being unsure
what the close button does.