Manual Drive better documented and Editor updated
Coded the buttons as to transmit motor to go while pressed and transmit
motor to stop when released. Couldn't figure out keyboard binding. Also
tried to code the communication between module and form as best I could.

Right and left turns involve having one motor turn forward and the other
one turn in reverse. Is that possible for the existing robot design?