Manual Drive module update
Added Manual Drive form, which contains the controls that were
previously in the editor and a set of bars and labels to provide motor
status for the user. Seems to be working, but I don't know how to make
it send data back to the module.

Modified Editor to contain priority and speed modifiers and check box to
enable dynamic drive.

Added variable for dynamic drive to IModule.