MotorStart and ManualDrive updates
Fixed error for when Dynamic Drive is null, logger no longer goes crazy
and now motorstart informs the logger if motors are on or off.

I made that algorithm in such a way that I don't have to go fiddling
with delegates and threading but it turned out kinda wonky... I'll try
and optimize it so it's more accurate.